show images from albums in admin list albums

+++ gallery/src/main/java/softuniGallery/controller/AlbumController.java
@@ -2,6 +2,7 @@ package softuniGallery.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -88,6 +89,15 @@ public class AlbumController {
             return "redirect:/album/viewAlbums";
         }
 
+        if (!(SecurityContextHolder.getContext().getAuthentication() instanceof AnonymousAuthenticationToken)) {
+            UserDetails principal = (UserDetails) SecurityContextHolder.getContext()
+                    .getAuthentication().getPrincipal();
+
+            User entityUser = this.userRepository.findByEmail(principal.getUsername());
+
+            model.addAttribute("user", entityUser);
+        }
+
         Album album = this.albumRepository.findOne(id);
 
         model.addAttribute("album", album);
@@ -105,6 +115,10 @@ public class AlbumController {
 
         Album album = this.albumRepository.findOne(id);
 
+        if (!isUserAuthorOrAdmin(album)) {
+            return "redirect:/album/" + id;
+        }
+
         model.addAttribute("view", "album/edit");
         model.addAttribute("album", album);
 
@@ -119,6 +133,10 @@ public class AlbumController {
         }
 
         Album album = this.albumRepository.findOne(id);
+
+        if (!isUserAuthorOrAdmin(album)) {
+            return "redirect:/album/" + id;
+        }
 
         album.setName(albumBindingModel.getName());
 
@@ -168,6 +186,10 @@ public class AlbumController {
 
         Album album = this.albumRepository.findOne(id);
 
+        if (!isUserAuthorOrAdmin(album)) {
+            return "redirect:/album/" + id;
+        }
+
         model.addAttribute("album", album);
         model.addAttribute("view", "album/delete");
 
@@ -183,6 +205,10 @@ public class AlbumController {
 
         Album album = this.albumRepository.findOne(id);
 
+        if (!isUserAuthorOrAdmin(album)) {
+            return "redirect:/album" + id;
+        }
+
         List<String> imagesPath = album.getImagePathList();
 
         deleteFiles(imagesPath);
@@ -190,5 +216,14 @@ public class AlbumController {
         this.albumRepository.delete(album);
 
         return "redirect:/album/viewAlbums";
+    }
+
+    private boolean isUserAuthorOrAdmin(Album album) {
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+
+        User userEntity = this.userRepository.findByEmail(user.getUsername());
+
+        return userEntity.isAdmin() || userEntity.isAuthor(album);
     }
 }
